feat:: Title for all the blogs + new blog
