+++ website/pages/blog/announcing-cloudquery-terraform-drift-detection.md
@@ -6,9 +6,9 @@ description: Use CloudQuery to detect IaC drift locally and in Continuos Integra
 author: yevgenypats
 ---
 
-import { Authors } from "../../components/Authors"
+import { BlogHeader } from "../../components/BlogHeader"
 
-<Authors/>
+<BlogHeader/>
 
 
 **This feature was deprecated, [see blog post](https://www.cloudquery.io/blog/terraform-drift-deprecation).**
