+++ website/pages/blog/aws-improvements-Q1-2022.md
@@ -8,9 +8,9 @@ description: >-
 author: benjamin
 ---
 
-import { Authors } from "../../components/Authors"
+import { BlogHeader } from "../../components/BlogHeader"
 
-<Authors/>
+<BlogHeader/>
