+++ website/pages/blog/cloud-asset-inventory-cloudquery-apache-superset.md
@@ -11,9 +11,9 @@ description: >-
 author: yevgenypats
 ---
 
-import { Authors } from "../../components/Authors"
+import { BlogHeader } from "../../components/BlogHeader"
 
-<Authors/>
+<BlogHeader/>
 
 
 In this blog post, we will walk you through how to setup CloudQuery to build your cloud asset inventory in PostgreSQL and connect it to [Apache Superset](https://superset.apache.org/) (or a hosted version such as [preset.io](https://preset.io)) for visualization, monitoring and reporting.
