+++ website/pages/blog/cloudquery-postgrest.md
@@ -6,9 +6,9 @@ description: See everything you have in the cloud with PostgREST
 author: yevgenypats
 ---
 
-import { Authors } from "../../components/Authors"
+import { BlogHeader } from "../../components/BlogHeader"
 
-<Authors/>
+<BlogHeader/>
 
 
 In this blog post, we will walk you through how to set up [CloudQuery](https://github.com/cloudquery/cloudquery) to build your cloud asset inventory in PostgreSQL and build a fully automated Restful API query layer with [PostgREST](https://postgrest.org/en/stable/). This can be used as a base for many custom use cases: from infrastructure search to security, cost and infrastructure automation.
