+++ website/pages/blog/releasing-cloudquery-0-20-0.md
@@ -8,9 +8,9 @@ description: >-
 author: roneliahu
 ---
 
-import { Authors } from "../../components/Authors"
+import { BlogHeader } from "../../components/BlogHeader"
 
-<Authors/>
+<BlogHeader/>
